редагування назви
