+++ README.md
@@ -1,2 +1,2 @@
 # image-encoding
-шифрування монохромних зображень з використанням бінарних перетворень
+шифрування монохромних зображень з використанням бінарних та тернарних перетворень
